backend/requirements.txt
# Document Design Replicator - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Document Processing - DOCX
python-docx==1.1.0
mammoth==1.6.0

# Document Processing - PDF
PyMuPDF==1.23.8
pdfplumber==0.10.3
reportlab==4.0.8

# OCR
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# AI
openai==1.12.0

# Validation & Security
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Handling
# python-magic==0.4.27  # Optional: requires libmagic system library (brew install libmagic on macOS)
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.0
isort==5.13.2
flake8==7.0.0
